+++ R/check_pkgs.R
@@ -1,0 +1,4 @@
+check_pkgs <- function(pkgs){
+    if(is.character(pkgs)) pkgs <- data.table::data.table(package=pkgs)
+    return(pkgs)
+}